projet_data_integ-Final 3/population_survey_project/kafka_producer/requirements.txt
kafka-python==2.0.2
pyspark